Fix StringCommandParameter
Update common lib where the following change is applied: when injector tries to resolve StringCommandParamater, it raises error as it's unable to use the current constructor. Add StringParameterBuilder class that is used to generate string parameters. Add createMandatoryParameter method to it, so it will generate mandatory parameter with specified error message. Add IStringParameterBuilder interface.
The idea is to use stringParameter from injector resolver when the parameter will not be mandatory and to use the builder when a more complicated logic is needed.

Also fixed FileSystemStub as it had to implement new method from IFileSystem