Fix tabtab for -- options
Add implementation of IOptionsService interface with one method - getKnownOptions. This method is used in commands service in order to get full list of supported -- options. As each CLI have its own list of knownOptions and the mobile lib has only some of them, the implementation of the interface has to be outside of the lib. This way we can use autocomplete for all available options, not only for the common ones.