Add dashedOptions for each command
Each command can use specific dashed options (`-- avd <name>` for example is valid for one command, but not for other). Add dashedOptions parameter to ICommand interface and make sure to use it for validation when it exists. In case there's no dashedOptions paramter, use the common options. In order to validate the options, delete yargs from require and require it again as yargs.options(this.options).argv is caching the results and calling it more than once with different options is giving strange results.
Update yargs to 3.12.0