Error-reporting command
Add error-reporting command that should be used in a similar manner as feature-usage-tracking command. The difference is that during installation, the exceptions tracking will be enabled without prompt (keep current behavior that we are tracking exceptions). It can be disabled later by using `$ tns error-reporting disable`.
Refactor analytics-service in order to "hide" some methods from the interface and to simplify adding of new analytics commands and variables.

**Rename feature-usage-tracking to usage-reporting**