Add platform on all commands
When you execute `tns run <platform>`, but you have forgotten to add the platform before that, you go in inconsistent state. Make sure to add the platform for the user when one of the following is executed:
* `tns prepare <platform>`
* `tns build <platform>`
* `tns run <platform>`
* `tns update <platform>@<version>` - this will add the specified version in case you have not added the platform before that.
* `tns emulate <platform>`

This way we simplify the workflow, now you can just do:
1) `tns create app1`
2) `tns run android --path app1`
And the application will run on your attached device.

Resolves https://github.com/NativeScript/nativescript-cli/issues/774
Part of the implementation of: https://github.com/NativeScript/nativescript-cli/issues/475