Add --copy-to option to build command
Add `--copy-to` option to build command that allows copying the built file (.apk or .ipa)
to a specified place. For example:
```
tns build android --copy-to ./
```
or
```
tns build android --copy-to ./myDir/out.apk
```
In case the output dir does not exist, it will be created. In case the specified value
for `copy-to` option is directory, the original filename will be used.

Required for easier integration of NativeScript in AppBuilder.