Do not call npm install if everything is installed
In case there's node_modules dir and there's directory for each dependency defined
in package.json of the project, we should not call `npm install`. In case `node_modules` dir
does not exist and the project has dependencies or any of the dependencies is not installed,
we'll call `npm install`. Another case when we'll call `npm install` is when `--force` option is used.

This change is in order to allow modifications of modules inside `node_modules` directory.