MLIBZ-667 Silence database errors during init
When the library is initialized, it will perform an upgrade operation on the database. NodeJS doesn't have a concept of a database resulting in a methodNotSupported error being thrown. This error would be logged and shown to the user causing concern when in fact nothing had gone wrong in the SDK. The error is now silenced.

Signed-off-by: Thomas Conner <thomas@kinvey.com>