Fix run on multiple devices
When multiple devices with same platform are attached, `tns run <platform>` command should start the app on all of them.
Currently the application is started, but the call to commandsService (device|run) command is failing as more than one device is attached.
Fix this by calling the correct method from the devicesService.