Use fork instead of spawn for startTestRunner
In `karma-nativescript-launcher` we call startTestRunner through spawn.
Communication between karma launcher and tns is through stdin.
This is causing exception in buildMetadata step when gradle wrapper is used.
For some reason it throws error in java code.
So, replace the spawn with fork in the launcher. Modify `startTestRunner` to wait for message from parent process and send "ready" state to the parent.
After that the parent (karma-nativescript-launcher) sends configuration data and the `startTestRunner` will execute the actions.

Fixes https://github.com/NativeScript/nativescript-cli/issues/1199