Add help for dev-test-* commands
As the current workflow of unit-testing is:
1) user calls `tns test android`
2) CLI starts a new process in which `tns dev-test android` is called
3) In case the dev-test command fails, the error is caught in the child process and command help is printed on the terminal directly in it.
4) User receives message: `Unable to find command dev-test|android` (from the child process) as we do not have help for this commands.

Copy-paste the help fo test-android and test-ios commands and change the filenames so when the child process fails, it will print correct help.
Parent process will not fail as the error is caught inside the child, so there's no way to print the help twice.

Update common-lib to search dev-commands as well when searching for help pages.