Speed up livesync on android device by using hash file
Optimize the usage of `tns livesync android` command. In case your project is huge, every time you use it, the full project will be transferred to device.
Persist all file hashes of project's files and next time when command is executed transfer only the changed ones.
The json which contains hashes information should be on device, so when the device is attached to another PC, where the same project is developed, we'll still transfer only the changed files.
Related:  https://github.com/NativeScript/nativescript-cli/issues/921