Speed-up grunt lint
Currently `grunt lint` might take more than 2-3 minutes. The problem is in the src files passed.
In my case the scratch directory was around 2.5 GB and this was slowing down the linting, as we have included `!*.d.ts` files.
Make sure only lib and test directories are used when linting as our source code and tests are there.