Use AndroidManifest from App_Resources as full manifest
For iOS we'll have full Info.plist in app/App_Resources/iOS.
Currently for Android we allow placing a file, that will be merged with default AndroidManifest from the runtime.
Change the Android implementation, so now the AndroidManifest placed in app/App_Resources/Android will be the full one.
In such file is not present in user's project, extract it from default template.
In case there's already such file and it doesn't seems correct, rename it to old.AndroidManifest.xml, extract correct manifest from template and inform the user that he'll have to merge the files manually.

Validate the AndroidManifest.xml

Validate the AndroidManifest.xml in app/App_Resources/Android. In order to achieve this, introduce new XmlValidator class.