Add support for custom platform templates
Add new option when adding platform: `--platform-template`. When it is
used, CLI will use the specified template instead of the default template
from the runtime. The path to the specified template will be saved in
project's package.json, so next time when platform is added, the same
template will be used.
In case when `--platform-template` is not passed, CLI will check the
package.json and if there's value for the template, it will be used.
Otherwise the default template from runtime will be used.