Fix starting of emulator when no devices are attached
When there are no devices attached, CLI should start emulator and execute the action on it.
For example calling `tns run android` when there are no devices attached, the devicesService should start emulator and run the application in it.

Currently this is not working as the startEmulator method is starting the instance and then calls method for device detection.
However the checkForDevices method is async and it does not guarantee when the devices will be detected.
Instead use startLookingForDevices which is synchronous.