Replace plugin variables in all Android .xml files
In case a plugin has plugin variable, during project preparation CLI replaces the usage of the variable in plugin's AndroidManifest with the value set in the current project's package.json.
This is not enough, we have to set the variables in all plugin's .xml files, so the value could be used in resource and later read from the JAVA code.
Enumarate plugin's files and replace plugin variables in all .xml files.