Added check if project name starts with number when creating new project.
Created project name service with method to ensure the project name does not start with number or if it starts with number the client is informed about the consequences when building for Android.
When invalid project name is entered the client gets warning about the problem and a prompt with choice to create the project with invalid name or to enter valid name.
Added integration tests for the prompts logic.
Added unit tests for the project name service.