Add devDependencies when adding Android platform
In order to support Static Binding Generation, we have to add some dev Dependencies to the project when platform Android is added.
As some of them do not work with Nodejs 0.x, we have to limit this functionality only to users who are using node 4.2.1 (the earliest version of 4.x that we support) or later.
Show warning in case the user has older node version.
Remove --staticBindings flag as it's turned on by default now.