Hide npm preuninstall errors
Due to issue in npm versions 3.0, 3.5.1, 3.7.3 (may be more 3.x versions) the NativeScript node_modules are removed before the preuninstall script is executed and it fails.
The issue is logged in the npm repository: https://github.com/npm/npm/issues/8806 but it is not fixed in version 3.1.1 as commented in the discusion.