Update dependencies in order to fix minimatch warning
As some of the dependencies are using old version of minimatch, the following warning appears during installation of nativescript CLI:
`npm WARN deprecated minimatch@0.2.14: Please update to minimatch 3.0.2 or higher to avoid a RegExp DoS issue`

Update some of the dependencies in order to stop showing the warning. However it is still there as the broccoli dependency is not updated for more than 8 months and it depends on old version of glob module.
Same is valid for the node-inspector version that we are using and that we cannot upgrade at the moment.