Fix first unit tests execution
First call to `tns test <platform>` always fails as the file with Karma configuration is not sent to device. The reason is that it's craeted in `node_modules/unit-test-runner` after the project is prepared.
The second execution of the command will include the file, but it may not be correct. It's regenerated, but again after the project is prepared.
Fix the `test-execution-service` to prepare the project after the configuration file is created.
Fixes https://github.com/NativeScript/nativescript-cli/issues/1636 and possibly https://github.com/NativeScript/nativescript-cli/issues/2016
Fixes the problem in Travis builds that console hangs and travis builds never end.