Fix failing android build command by forcing use of gradle
When there's `platforms/android` directory, but in package.json there's no entry in `nativescript` for android platform, calling `tns build android` or any build related command will fail with error:
`Cannot read property 'version' of undefined`
The reason is that we are trying to check the version in package.json and check if it allows us to use gradle (previously Android builds required Ant).
Prevent the error and force using gradle in such case.