Add support for Node.js 6.5 (v8 5.1)
Add support for Node.js 6.5 (v8 5.1) by rebuilding our native (C++) dependencies and using the updated versions in the package.json.
Node.js 6.5 has new version of the v8 - from 5.0.x to 5.1.x. When rebuilding and requiring our native dependencies, we are using the version of v8 in order to find correct binary.