Get latest common lib
Get latest common lib with two fixes:

We need to check if there are any emulators installed for android to emulate on android.

Use the monkey tool to start Android applications on device. This way we do not have to care about the startPackageActivity name and we can start the application in case a custom activity is used.
Remove unused code and variables.