Stop calling npm install on each prepare of ng projects
During project preparation CLI's trying to detect if all dependencies from package.json are installed by listing contents of node_modules directory.
In case any of them are missing, CLI calls `npm install` inside the project. This executes all postinstall scripts of the dependencies and leads to slower prepare (and also slower builds in lots of cases).
The problem is that the code is reading the contents of node_modules directory, but scoped dependencies (used in all angular projects) are not in the root of node_modules.
So in all Angular projects each prepare calls `npm install`. Fix the listing of the directories to include scoped packages, so in case everything is installed, CLI will not call npm install on the next preparation of the project.