Do not stop process when EPEERINVALID error is raised
In many cases `npm install` command, that CLI calls, will raise EPEERINVALID error (on Linux and Mac OS X).
This error code indicates that some of the dependencies have peer dependencies that are not included in the project.
Npm 2 considers this case as error and raises EPEERINVALID error. However in npm 3 this case is treated as non-errornous case and instead of EPEERINVALID, only WARNINGIS are shown.
CLI uses npm 2 internally (as dependency), so whenever such case happens, it breaks its execution.
Until the dependency is updated, let's treat these errors as warnings and continue the execution. After we upgrade the dependency to npm 3 we'll have the same behavior.