Show helpful information when Sync over WIFI error is detected
iTunes has an option to sync changes over wifi for each iOS device. However, for some devices, when trying to start the installation_proxy service and `Sync over Wi-Fi` option is enabled, the connection fails.
Catch such cases and show helpful message describing the problem and how to resolve it.