Merge post_install hooks in Podfile
Currently the CLI does not merge the post_install hooks in the Podfile. When there are more than one post_install hooks cocoapods will fail on pod install. The solution is to find all the post_install hooks in the merged Podfile and transform them to methods. After that post_install will be called with block in which all the transformed methods will be called.