Use https for submodule initialization
At the moment the submodule is initialized via ssh. This way the command `git submodule update --init` will work only for users who have write access to mobile-cli-lib.
This also requires custom logic when travis is used. In current implementation, travis automatically calls `npm install` before executing the scripts, but at this point the submodule is not initialized yet (as we have declared submodules variable to false).
So the `npm install` executed by travis, will fail in case the postinstall script uses some code from the submodule.
In order to fix this two changes are applied:
 - change the sumodule to be used over https instead of ssh.
 - modify .travis.yml to use submodule correctly and remove two manual steps which are executed by travis by default, there's no need the execute them again.

NOTE: In case you have two-factor authentication, you'll not be able to commit to mobile-cli-lib over https. In this case you have to:
 - remove content of lib/common dir
 - clone over ssh: `$ git clone git@github.com:telerik/mobile-cli-lib.git .` (you must be inside lib/common dir).