Use sync version of fs.exists
We use fs.exists, which is deprecated. Change the code to use fs.existsSync instead. This makes the `$fs.exists` method to return boolean instead of `IFuture<boolean>`.
Several methods were declared as `IFuture<T>` as in their body they are calling `$fs.exists(...).wait()`. As now the `$fs.exists` method is sync, change all methods that are async only because of `$fs.exists().wait()` sync.