Expose createProject to public API
In order to allow creating a new NativeScript project when {N} CLI is required as a library, we should add createProject method to public API. This requires some modifications in method's implementation - we cannot rely on `$options` as it is not populated correctly when the CLI is used as a library. Pass object describing the project that we have to create. Fix unit tests accordingly and uncomment (and fix) all tests in project-service.

Add unit tests that will check if a method is really exposed in public API.

Add PublicAPI.md file, which describes all methods that can be used when using CLI as a library.