Add tracking for the project types that users are creating and working with (#2497)
* Track from which template a project is created

Add tracking from which template a project is created. This will give us better picture of the usage of CLI and the types of projects that the users are creating.

* Track project type when deploy/run/livesync is executed

Track the project type (Angular, Pure TypeScript, Pure JavaScript) when any of the commands is executed:
* prepare
* deploy
* run
* livesync

This will allow us to better understand the type of projects that the users are building.