Pass current stdio to npm install process (#2527)
Whenever CLI spawns `npm install <smth>`, we do not set stdin and stdout
of the spawned process. This way in case any of the npm scripts executed
during install requires input from user, it cannot be handled in any
way.

So pass the current stdin and stdout to the spawned npm process. This
way in case a postinstall script is executed and it requires input from
user, the users will be able to enter required information in the same
place where `tns` command has been executed (same terminal).

This fixes postinstall execution of `nativescript-plugin-firebase`.