Merge LiveSync fixes from release branch (#2553)
* Fix uploading of .nsprepareinfo on Windows

When `.nprepareinfo` file is uploaded on device, we construct the device file path with `path.join` function. On Windows this function returns path similar to `\\data\\local\\tmp`, which is not working for Unix based OSes (Android in this case).
Fix this by converting the path to Unix style.

* Fix uploading of .nsbuilinfo on Windows

When `.npbuildnfo` file is uploaded on device, we construct the device file path with `path.join` function. On Windows this function returns path similar to `\\data\\local\\tmp`, which is not working for Unix based OSes (Android in this case).
Fix this by converting the path to Unix style.

* Fix ENOENT errors on macOS during livesync

On macOS sometimes we recive ENOENT errors during livesync operations (`tns run ios`). This is due to the following:
- incorrect detection if a file is modified - we check if the project dir's mTime and cTime values are bigger than the .nsprepareinfo. However in some cases they are equal, which means the dir has been modified in the same moment when the `.nsprepareinfo` is saved. In order to fix this, mark the files/dirs as modified in case the modified time of `.nsprepareinfo` and the respective file are the same.
- In some cases we are not able to sync files to iOS Simulator as chokidar does not raise correct events on macOS when directory is renamed immediately after it's being created.
After adding a file to this directory we throw ENOENT, as such dir does not exist in Simulator's dir. In order to fix this, ensure the directory exist in the simulator.

PR in common-lib: https://github.com/telerik/mobile-cli-lib/pull/892

* Update to latest common lib