Fix creating multiple projects in a single process
In case the CLI is `required` as a library (used in a long living process, not as a command-line interface), trying to create a project multiple times leads ot error.
The problem is in the `projectDataService` which caches the path to first created project file (package.json) and uses it for all consecutive operations.
In other cases, when CLI had to use the projectDataService, it always had to call `projectDataService.initialize` method.

Instead of relying on a magically called initialize method, modify the service methods to work with any project dir and execute the required operation for it.
This will fix the creation of multiple projects as all operations will work with currently passed project directory.
Also remove the code that creates a `package.json` file during `loadProjectFile` in projectDataService - it's required only when a project is created, so move it there.

Rename the methods getValue, setValue and removeProperty to getNSValue, setNSValue, removeNSProperty as they are intended to modify the nativescript key in package.json.
We can later expose getValue, setValue and removeProperty (they are currently private).

Remove the indentation detection as `$fs.writeJson` file already does this.