Fix adding of invalid platform
When user tries to add invalid platform, CLI should prints user-friendly
message. Instead it fails that cannot read `getPlatformData` of
`undefined`. The reason is a change in platforms-data `getPlatformsData`
method. Fix it to return undefined when there's no platform specific
data. This way the code will detect the invalid platform and will fail
with user-friendly message.