Pass android specific options for emulate command
In case you want to use `tns emulate android --release`, several android
specific options should be passed and used by the service. Even when
they are passed on the command line, we did not pass them to the
service. Pass required options, so the release build will be
successfully signed.