Fix several issues in `@next` version (#2591)
* Fix adding of invalid platform

When user tries to add invalid platform, CLI should prints user-friendly
message. Instead it fails that cannot read `getPlatformData` of
`undefined`. The reason is a change in platforms-data `getPlatformsData`
method. Fix it to return undefined when there's no platform specific
data. This way the code will detect the invalid platform and will fail
with user-friendly message.

* Fix specifiying --sdk on platform add

During adding of Android platform, users can specify targetSdk via
`--sdk` command line option. This has been broken in previous commit.
As our services should not rely on `$options`, sdk must be passed as
parameter whenever it could be used. It turns out some common methods
(addPlatform, preparePlatform) rely on platform specific data - sdk,
provision, etc. In order to handle this without adding new arguments,
introudce IPlatformSpecificData interface and use it in al methods
where we have to pass some of the platform specific data required for
preparing of the project.

* Pass android specific options for emulate command

In case you want to use `tns emulate android --release`, several android
specific options should be passed and used by the service. Even when
they are passed on the command line, we did not pass them to the
service. Pass required options, so the release build will be
successfully signed.

* Fix livesync watch operations

LiveSync provider that detects how to process the files, needs the
projectData. However we had not passed it and the code fails. Pass the
data wherever its needed. This fixes livesync operations.