Remove ios native dependencies (#2495)
* Remove ios native dependencies

We need to remove our native dependencies - ref and ffi because they are casuing problems with each version of NodeJS which has new version of V8 in it. We are using ref and ffi to communicate with iTunes and execute fs and application management operations on iOS devices. This operations are moved in new dependency - device-lib. The device lib has JavaScript wrapper which we use in the CLI. Everything in the ios device file system and ios device application manager is replaced with the exposed methods from the device lib.

* Change the notification logic

* Post-rebase fixes

* Add receive message from ios device socket logic

* Fix rebase issues

* WIP

* Fix tns debug ios issues

Since we use the ios-device-lib to connect to the device backend port for debugging, the socket returned from the device can be used only in the ios-device lib.
That's why we need to change the logic for creating the Node.js socket in the CLI. The ios-device-lib will expose the device socket on a specific port in the localhost.
After we connect to the exposed socket we can continue to work like we used to.

* Update ios-devic-lib version